[feat] 가계부 내역 코드 구조 변화
